Convert GameBox to TS
